+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/packet/c2s/ClientToServerCommunicationsImpl.java
@@ -88,7 +88,7 @@ public class ClientToServerCommunicationsImpl implements ClientToServerCommunica
     }
 
     @Override
-    public void sendDetectorAmountChange(final long amount) {
+    public void sendDetectorAmountChange(final double amount) {
         networkManager.send(new DetectorAmountChangePacket(amount));
     }
 }
